requirements.txt
django
django-cors-headers
djangorestframework-jwt
djangorestframework
django-log-request-id
django-ratelimit==1.1.0
